// electron-app/src/server/settings/settings.interface.ts
export interface Settings {
  advertise: boolean;
  emptyQueueOnFailedPost: boolean;
  postRetries: number;
  openOnLogin: boolean;
  openWindowOnStartup: boolean;
  useHardwareAcceleration: boolean;
  maxPNGSizeCompression: number;
  maxPNGSizeCompressionWithAlpha: number;
  maxJPEGQualityCompression: number;
  maxJPEGSizeCompression: number;
  silentNotification: boolean;
}